+++ packages/relax_cli/lib/src/templates/bloc_app_template.dart
@@ -98,7 +98,7 @@ dependencies:
   go_router: ^14.6.0
   slang: ^4.14.0
   slang_flutter: ^4.14.0
-  relax_orm: ^0.1.4
+  relax_orm: ^1.0.0
   relax_storage: ^1.0.1
   env:
     path: packages/env
@@ -108,7 +108,7 @@ dev_dependencies:
     sdk: flutter
   flutter_lints: ^5.0.0
   build_runner: ^2.4.0
-  relax_orm_generator: ^0.1.6
+  relax_orm_generator: ^0.1.7
 
 flutter:
   uses-material-design: true