+++ packages/relax_cli/lib/src/templates/getx_app_template.dart
@@ -83,7 +83,7 @@ dependencies:
   get_it: ^8.0.3
   slang: ^4.14.0
   slang_flutter: ^4.14.0
-  relax_orm: ^0.1.4
+  relax_orm: ^1.0.0
   relax_storage: ^1.0.1
   env:
     path: packages/env
@@ -96,7 +96,7 @@ dev_dependencies:
   
   flutter_lints: ^5.0.0
   build_runner: ^2.4.0
-  relax_orm_generator: ^0.1.6
+  relax_orm_generator: ^0.1.7
 
 flutter:
   uses-material-design: true